Try to get the file name from the content-type, not the content-id
In a PNG file I was texted, the content-id didn't have the file name as I expected.  Let's try to get it from the content-type first, and fallback to the content-id if we need.

Does the content-type always contain a charset?  I'm reading 0x85 as one, but this may just be "start of file name".  For images, this makes sense, but does the SMIL file need a charset?

Sometimes the 0x85 byte is included after the charset, and sometimes it *is* the charset (I think).  So, I guess some content-types (like SMIL) have implicit charsets?